Add some shortcuts closes #72
Now, PDF4Teachers can remembre the scroll location when closing a file. This only works with documents with an edition, you do not have to save the file to save the scroll location (This is done automatically when there is an edit file).

I also added a lot of shortcuts:
- Ctrl + Backspace is considered ad Delete (For non-full keyboards and laptop)
- TAB can still be used to select the text tab or now, the paint/images tab if the text tab is already selected.
- Page Up/Down support
- End/Beginning or End/Home support
- Arrows can be used to scroll a bit vertically or horizontally
- Ctrl + Up can be used to zoom a bit
- Ctrl + Down is fitting the pages automatically in the available width (equals to the new fit button in the footer bar).
- Ctrl + Alt + Up is equals to Page Up
- Ctrl + Alt + Down is equals to Page Down
- Ctrl + Alt + Right is loading the next file in the files list
- Ctrl + Alt + Left is loading the previous file in the files list

Btw, the zoom animations options is now divided in two settings: trackpad mode and animations, so you can still have the animations other than the trackpad zoom/scroll while being on trackpad.