Support undo for skills
Changing the assessment linked to the document does not count as an UndoAction, but when a notation change is undone, it will automatically switch on the right assessment.
Student selection is not managed with UndoAction. When switching assessment, the selected student could be unselected if not present in the new assessment students list.

Took 41 minutes