Fix issue with grade choose value menu staying open when closing parent menu.
Grade choose value menu is not anymore a part of PageRenderer because it is not only generated in PageRenderer. (Also on the elements or on a sidebar).
The parent menu still stays open, but when one of the two menus is closed, the second one is automatically closed too.

Took 1 hour 11 minutes