New shortcuts organization for #166
The app shortcuts (except the menubar shortcuts) are now represented as an entry with the KeyCombination and a Consumer of KeyEvent.

This new organization will allow to be able to check what shortcuts are already used by the app or not for #166.

An issue in JavaFX makes MacOS non QWERTY keyboards not working well with KeyCodes that returns the key value on a QWERTY keyboard. This issue should be fixed in JFX 21 that is not released yet.

Took 8 minutes