Fixing scroll memorization issue + Keeping the same zoom factor when switching between documents closing #170
The last scroll value is now properly restored (was not always working when switching between documents). This was fixed by adding a Platform.runLater() after closing the last document and before opening the new one. This simulates a clean document closing process.

As detailed in #170, the default zoom factor is not anymore always dynamically calculated to fit the document. If a document was opened when opening a new document, the new document will heritate it's default zoom factor from the last one. (The zoom factor don't changes when switching between documents)