FEAT: allow to dran'n drop files/images to open on the MainScreen.
If releasing the drag on a page with the dragboard containing only images/svg, the image/svg will be added as an element instead of converting.

The left tab is selected only if more than one file is dragged, and the first dragged file will always be opened as the current document.
When dropping in the files sidebar, a document will be opened only if it is the only dragged file (unchanged).