FIX: Big documents gets corrupted when editing document
Editing a document (page rotation, booklets, deleting pages) corrupted some pages of the document because PDF4Teachers was overwriting the currently opened file, which led to corrupted pages.

This commit fixes this issue making sure at each document save that it is not the current file, and if it is, it will save the new file as a temp file, and move it afterward.

This commit also adds a better behavior of the PDFPagesRender.